fix circular input
